+++ src/stock-trading-vue/static/components/stock/Stock.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { onMounted, inject, reactive, ref } from 'vue';
+import { inject, ref, computed, watchEffect } from 'vue';
 import { AxiosInstance } from 'axios';
 import { httpClientKey } from '../../plugins/http_client'
 import { StockListResponse } from '../../models';
@@ -7,20 +7,27 @@ import { StockListResponse } from '../../models';
 const httpClient = inject<AxiosInstance>(httpClientKey) as AxiosInstance
 const stockListResponse = ref<StockListResponse>()
 
-onMounted(() => {
-    listStocks()
+const currentPage = ref<number>(1)
+const pageCount = computed(() => {
+    if (stockListResponse.value) {
+        return Math.floor(stockListResponse.value?.count / import.meta.env.VITE_API_PAGE_SIZE)
+            + ((stockListResponse.value?.count % import.meta.env.VITE_API_PAGE_SIZE)? 1:0)
+    }
+    return 0
 })
+const changePage = (page) => currentPage.value = page
 
-function listStocks() {
-    httpClient.get('/stocks')
+watchEffect (async () => {
+    let url = `/stocks?page=${currentPage.value}`
+    stockListResponse.value = await httpClient.get(url)
         .then(function (response) {
-            stockListResponse.value = response.data as StockListResponse
-            console.log(stockListResponse.value)
+            return response.data
         })
         .catch(function (error) {
-            console.log(error);
+            // TODO: refresh token
+            console.log(error)
         })
-}
+})
 </script>
 
 <template>
@@ -46,4 +53,22 @@ function listStocks() {
             </tbody>
         </table>
     </div>
+
+    <nav aria-label="Page navigation example">
+        <ul class="pagination">
+            <li class="page-item" :class="[stockListResponse?.previous?'':'disabled']">
+                <a class="page-link" href="#" aria-label="Previous">
+                    <span aria-hidden="true">&laquo;</span>
+                </a>
+            </li>
+            <li v-for="p in pageCount" :key="p" class="page-item" :class="[currentPage==p?'disabled':'']">
+                <a class="page-link" href="#" @click="changePage(p)">{{ p }}</a>
+            </li>
+            <li class="page-item" :class="[stockListResponse?.next?'':'disabled']">
+                <a class="page-link" href="#" aria-label="Next">
+                    <span aria-hidden="true">&raquo;</span>
+                </a>
+            </li>
+        </ul>
+    </nav>
 </template>