+++ CHANGELOG.md
@@ -8,6 +8,11 @@
 - Trade Page
 - Portfolio Page
 
+[0.0.10] - 2024-07-12
+### Added
+- Stock list pagination
+- Implement Vue WatcherEffect
+
 [0.0.9] - 2024-07-12
 ### Added
 - Stock list